feat: remove getrandom feature list for wasm
It's better to allow the upstream consumer of this library to specify
how they want to manage wasm compilation. For example I'm using the
"custom" feature of getrandom instead of the "js' feature. Having it
specified in this library enforces "js" feature which makes getrandom
inconsistent across my wasm project, some is custom, some is js.